requirements.txt
beautifulsoup4==4.11.1
pandas==1.4.1
Pillow==9.1.0
streamlit==1.7.0
urllib3==1.26.8
click==7.1.2
PyMovieDb